backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
black==25.1.0
bs4==0.0.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
h11==0.16.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
Markdown==3.8
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
